Research/attacker/requirements.txt
prompt_toolkit>=3.0.0
pycryptodome>=3.15.0
dnslib>=0.9.7